+++ js/components/date_selector.js
@@ -1,6 +1,5 @@
 import Vue from 'vue'
 import { getDaysInMonth } from 'date-fns'
-import { emitEvent } from '../lib/emitters'
 
 let paddedNumber = function(number) {
   if ((number + '').length === 1) {
@@ -35,14 +34,6 @@ export default {
       year: this.initialyear,
       name: this.nameTag,
     }
-  },
-
-  created: function() {
-    emitEvent('field-mount', this, {
-      optional: this.optional,
-      name: this.name,
-      valid: this.isDateValid,
-    })
   },
 
   watch: {
@@ -217,13 +208,7 @@ export default {
   },
 
   methods: {
-    onInput: function(e) {
-      emitEvent('field-change', this, {
-        value: this.formattedDate,
-        name: this.name,
-        valid: this.isDateValid,
-      })
-
+    onInput: function() {
       this.$parent.$emit('field-change', {
         value: this.formattedDate,
         name: this.name,
