+++ js/components/forms/multi_step_modal_form.js
@@ -1,4 +1,4 @@
-import FormMixin from '../../mixins/form_mixin'
+import FormMixin from '../../mixins/form'
 import textinput from '../text_input'
 import optionsinput from '../options_input'
 import checkboxinput from '../checkbox_input'
